validaciones

+++ app/Controllers/CursosController.php
@@ -10,11 +10,12 @@ class CursosController extends BaseController
 {
     public function index(){
         $mcursos = new Cursos();
-        $data['cursos'] = $mcursos->findAll();
+        $data['cursos'] = $mcursos->paginate(10);
+        $paginador = $mcursos->pager;
+        $data['paginador'] = $paginador;
 
         return view('cursos/index', $data);
     }
-
 
     public function register(): string
     {
@@ -84,27 +85,33 @@ class CursosController extends BaseController
     public function eliminar($id){
         $mcursos = new Cursos();
         $datos_curso = $mcursos->find($id);
-
+        $inscripciones = new Inscripciones();
         if($datos_curso){
-            $mcursos->delete($id);
-            return redirect()->back()->with('success', 'Curso Eliminado');
-        }
-        else {
-            return redirect()->back()->with('error', 'Curso no Eliminado');
+            if($inscripciones->verificarInscripcionCurso($id) > 0){
+                return redirect()->back()->with('error', 'Curso no Eliminado');
+            }
+            else {
+                $mcursos->delete($id);
+                return redirect()->back()->with('success', 'Curso Eliminado');
+            }
+        }else{
+            return redirect()->back()->with('error', 'Curso no Encontrado');
         }
     }
 
+
     public function ver($id){
-        $curso = new Cursos();
-        $data['curso'] = $curso->find($id);
+        $mcurso = new Cursos();
+        $data['curso'] = $mcurso->find($id);
+
         $inscripciones = new Inscripciones();
         $data['inscripciones'] = $inscripciones->where('cursos_id', $id)->mostrar($inscripciones);
-        
+        $paginador = $inscripciones->pager;
+        $data['paginador'] = $paginador;
 
-        if ($data) {
+        if ($data) { 
             return view('cursos/show', $data);
-        } 
-        else {
+        } else {
             return redirect()->to(base_url()."cursos");
         }
     }
@@ -138,7 +145,7 @@ class CursosController extends BaseController
                     'estado' => 1,
                 ]
             ); 
-            return redirect()->to(base_url()."ver-cursos/".$id)->with('success', 'Curso creado');
+            return redirect()->to(base_url()."ver_inscripciones/".$id)->with('success', 'Curso creado');
         }
         return  redirect()->to(base_url()."inscribir-cursos/".$id)->with('error', 'La validación de datos falló. Por favor, revisa tus entradas:<br>
                                                                             - Estudiante (obligatorio)');
@@ -173,6 +180,19 @@ class CursosController extends BaseController
             return redirect()->to(base_url()."ver-cursos/".$inscripcion['cursos_id'])->with('success', 'Curso creado');
         } else {
             return redirect()->to(base_url()."cursos");
+        }
+    }
+
+    public function eliminar_inscripcion($id){
+        $minscripcion = new Inscripciones();
+        $inscripcion = $minscripcion->find($id);
+
+        if($inscripcion){
+            $minscripcion->delete($id);
+            return redirect()->back()->with('success', 'Inscripcion Eliminada');
+        }
+        else {
+            return redirect()->back()->with('error', 'Inscripcion no Eliminada');
         }
     }
 }
